requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
autopep8==2.3.2
aws-requests-auth==0.4.3
backcall==0.2.0
beautifulsoup4==4.14.2
bedrock-agentcore==1.0.6
bedrock-agentcore-starter-toolkit==0.1.28
bleach==6.3.0
boto3==1.40.64
botocore==1.40.64
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
executing==2.2.1
fastjsonschema==2.21.2
frozenlist==1.8.0
google-auth==2.42.1
google-genai==1.47.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
langchain-core==1.0.5
langchain-valyu==0.3.4
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.20.0
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
openai==2.8.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pathable==0.4.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.5.0
playwright==1.55.0
prance==25.4.8.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-openapi-schema-to-json-schema==0.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
questionary==2.1.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.26.0
slack_sdk==3.37.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
strands-agents==1.14.0
strands-agents-tools==0.2.13
sympy==1.14.0
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
valyu==2.2.2
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0